+++ packages/client/src/lib/index.ts
@@ -1,3 +1,6 @@
 export * from './room/index.js';
+export * from './registration/index.js';
+export * from './conversation/index.js';
+export * from './conversationList/index.js';
 export * from './API.js';
 export { originPublicKeyPublicationMessage } from 'missiv-common';