+++ platforms/cf-worker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # missiv-server-cf-worker
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - missiv-server@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes